Renamed functionality of TaskService to TaskRepository
