+++ RoadGap.webapi/Controllers/TaskController.cs
@@ -1,7 +1,7 @@
 using AutoMapper;
 using Microsoft.AspNetCore.Mvc;
 using RoadGap.webapi.Dtos;
-using RoadGap.webapi.Service;
+using RoadGap.webapi.Repositories;
 using Task = RoadGap.webapi.Models.Task;
 
 namespace RoadGap.webapi.Controllers;
@@ -10,12 +10,12 @@ namespace RoadGap.webapi.Controllers;
 [Route("tasks")]
 public class TaskController : ControllerBase
 {
-    private readonly ITaskService _taskService;
+    private readonly ITaskRepository _taskRepository;
     private readonly IMapper _mapper;
 
-    public TaskController(ITaskService taskService)
+    public TaskController(ITaskRepository taskRepository)
     {
-        _taskService = taskService;
+        _taskRepository = taskRepository;
         _mapper = new Mapper(new MapperConfiguration(configurationExpression =>
         {
             configurationExpression.CreateMap<TaskToUpsertDto, Task>();
@@ -25,14 +25,14 @@ public class TaskController : ControllerBase
     [HttpGet]
     public IActionResult Get()
     {
-        var tasks = _taskService.GetTasks();
+        var tasks = _taskRepository.GetTasks();
         return Ok(tasks);
     }
 
     [HttpGet("{taskId:int}")]
     public IActionResult Get(int taskId)
     {
-        var task = _taskService.GetTaskById(taskId);
+        var task = _taskRepository.GetTaskById(taskId);
         if (task == null)
         {
             return NotFound("There's no task with this id.");
@@ -44,7 +44,7 @@ public class TaskController : ControllerBase
     [HttpGet("{searchParam}")]
     public IActionResult Get(string searchParam)
     {
-        var searchedTasks = _taskService
+        var searchedTasks = _taskRepository
             .GetTasksBySearch(searchParam);
         
         return Ok(searchedTasks);
@@ -53,17 +53,17 @@ public class TaskController : ControllerBase
     [HttpPut("{taskId:int}")]
     public IActionResult Edit(int taskId, [FromBody] TaskToUpsertDto taskDto)
     {
-        if (!_taskService.CategoryExists(taskDto.CategoryId))
+        if (!_taskRepository.CategoryExists(taskDto.CategoryId))
         {
             return BadRequest("Invalid category id.");
         }
 
-        if (!_taskService.StatusExists(taskDto.StatusId))
+        if (!_taskRepository.StatusExists(taskDto.StatusId))
         {
             return BadRequest("Invalid status id.");
         }
 
-        var taskDb = _taskService.GetTaskById(taskId);
+        var taskDb = _taskRepository.GetTaskById(taskId);
 
         if (taskDb == null)
         {
@@ -72,7 +72,7 @@ public class TaskController : ControllerBase
 
         _mapper.Map(taskDto, taskDb);
 
-        _taskService.SaveChanges();
+        _taskRepository.SaveChanges();
 
         return Ok("Task updated successfully.");
     }
@@ -80,20 +80,20 @@ public class TaskController : ControllerBase
     [HttpPost]
     public IActionResult Create(TaskToUpsertDto taskToAdd)
     {
-        if (!_taskService.CategoryExists(taskToAdd.CategoryId))
+        if (!_taskRepository.CategoryExists(taskToAdd.CategoryId))
         {
             return BadRequest("Invalid category id.");
         }
 
-        if (!_taskService.StatusExists(taskToAdd.StatusId))
+        if (!_taskRepository.StatusExists(taskToAdd.StatusId))
         {
             return BadRequest("Invalid status id.");
         }
 
         var task = _mapper.Map<Task>(taskToAdd);
 
-        _taskService.AddEntity(task);
-        _taskService.SaveChanges();
+        _taskRepository.AddEntity(task);
+        _taskRepository.SaveChanges();
 
         return Ok("Task created successfully.");
     }
@@ -101,15 +101,15 @@ public class TaskController : ControllerBase
     [HttpDelete("{taskId:int}")]
     public IActionResult Delete(int taskId)
     {
-        var task = _taskService.GetTaskById(taskId);
+        var task = _taskRepository.GetTaskById(taskId);
         
         if (task == null)
         {
             return NotFound("There's no task with this id.");
         }
 
-        _taskService.RemoveEntity(task);
-        _taskService.SaveChanges();
+        _taskRepository.RemoveEntity(task);
+        _taskRepository.SaveChanges();
         
         return Ok("Task deleted successfully.");
     }
